Add monitoring of LSI disks and read only mounts
The C6220 servers come with an LSI controller that is not supported by
the OpenManage software that is used to monitor all the other Dell
machines. The patch adds a binary from LSI and a wrapper script to
integrate health checks on the disks and RAID volume into nagios.

There is also an added check for volumes that have gone read only. This
usually means that there has been filesystem corruption, but this
doesn't trigger any nagios failures. Sometimes I would get a failure
stating that puppet hadn't been run in a few hours, but that is not an
ideal detection mechanism.

Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>