Create mesos-agent whitelist using consul-template
Add the consul_template module and setup the service. The first template
is the whitelist which queries all the mesos-agents with a service that
is passing its health checks. This list is them pulled together and
automatically updated if a node is put into maintenance or the health
checks fail.

Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>