Add postmirror step to apt-mirror
Seems that InRelease file is not on official docker repos but apt-mirror
somehow creates it. But the created version causes a Hash Sum Mismatch
error on the clients. Turns out if this file is deleted all the problems
go away. This error has been reported to apt-mirror, but not fixed.