Expose dnsmasq on all interfaces
During puppet runs changes to dnsmasq caused lookup failures. Also if
docker was not installed dnsmasq server startup would also fail. Also
when docker compose/networking is used, the network is on a different
bridge which meant that networking would timeout and fail inside the
containers. This does expose dnsmasq to all traffic and is a potential
security issue.